Fixed the EXACTLY button

+++ week-06/day-5/basic-webshop/src/main/java/com/greenfoxacademy/basicwebshop/controllers/WebShopController.java
@@ -82,7 +82,7 @@ public class WebShopController {
 
   @PostMapping(path = "/search-by-price")
   public String searchByPrice(@RequestParam String searchMode,
-                              @RequestParam(defaultValue = "0") Integer number,
+                              @RequestParam(defaultValue = "0") float number,
                               Model model) {
     model.addAttribute("items", webShopService.getItemsByPrice(searchMode, number));
     return "morefilters";
